Initial support for advanced message uploading.
Now provides explicit control over message concurrency, support for upload
ordering with priority bypass (and exclusive priority slot to only have
latest message in the slot jump the queue).

Memory load if starting with many messages pending in the watched message
directory should be significantly reduced, as only metadata is kept until
the message is ready for publishing.

Also adding support for using a "letterhead", to allow for e.g. stamping each
message with a device-unique identifier, or a variable field such as a
batch number for the present manufacturing run.

Furthermore, the message payload may now be automatically compressed if so
requested.

Documentation yet to be written, together with additional tests.