Ensure outdated 'reported' section gets updated.
When receiving the full device shadow the "reported" section needs to be
checked to ensure it's matching our current reality, and updated if not.
I'd missed this check all this time because I had not gotten a device
shadow out of sync with the local state. Anyway, this finally addresses
this state-out-of-sync issue.